Hide Python imports from ipython
